arrglando problemas con tipos de datos

+++ iot/src/main/java/gt/com/iot/model/Events.java
@@ -27,7 +27,7 @@ public class Events implements Serializable{
     
     private Integer eventCode;
     
-    private Integer timeEvent;
+    private Long timeEvent;
     
     @ManyToOne
     @JoinColumn(name = "idSensor")
